# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic API"
  description: "This is an API to manage a Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint (generates token)"
      description: "Validates customer credentials and returns a customer-specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers:
    post:
      tags:
        - Customers
      summary: "Create/register a new customer."
      description: "Creates a new customer in the Mechanic API system. Emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomersPayload"
      responses:
        201:
          description: "Successful Customer Creation"
          schema:
            $ref: "#/definitions/CreateCustomersResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: johnd@email.com
              password: password123

    put:
      tags:
        - Customers
      summary: "Update a customer."
      description: "Token-authenticated route. Customer must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomersPayload"
      responses:
        200:
          description: "Successfully updated customer"
          schema:
            $ref: "#/definitions/UpdateCustomersResponse"
          examples:
            application/json:
              id: 1
              name: Dohn Joe
              email: dohnj@email.com
              password: password123

    delete:
      tags:
        - Customers
      summary: "Delete a customer"
      description: "Token-authenticated route. Customer must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully deleted customer"
          schema:
            $ref: "#/definitions/DeleteCustomersResponse"
          examples:
            application/json:
              message: "successfully deleted user <customer_id>."
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateCustomersPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - password
  CreateCustomersResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
...
  UpdateCustomersPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
        format: date
      password:
        type: "string"
    required:
      - name
      - email
      - password
  UpdateCustomersResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
  DeleteCustomersResponse:
    type: "object"
    properties:
      message:
        type: "string"
...